requirements.txt
pandas
requests
confluent-kafka
pymongo
python-decouple